Fix a couple warnings in the Units module (#1615)

+++ library/include/modules/Units.h
@@ -181,9 +181,9 @@ DFHACK_EXPORT std::string getCasteProfessionName(int race, int caste, df::profes
 DFHACK_EXPORT int8_t getProfessionColor(df::unit *unit, bool ignore_noble = false);
 DFHACK_EXPORT int8_t getCasteProfessionColor(int race, int caste, df::profession pid);
 
-DFHACK_EXPORT df::goal_type getGoalType(df::unit *unit, int goalIndex = 0);
-DFHACK_EXPORT std::string getGoalName(df::unit *unit, int goalIndex = 0);
-DFHACK_EXPORT bool isGoalAchieved(df::unit *unit, int goalIndex = 0);
+DFHACK_EXPORT df::goal_type getGoalType(df::unit *unit, size_t goalIndex = 0);
+DFHACK_EXPORT std::string getGoalName(df::unit *unit, size_t goalIndex = 0);
+DFHACK_EXPORT bool isGoalAchieved(df::unit *unit, size_t goalIndex = 0);
 
 DFHACK_EXPORT std::string getSquadName(df::unit *unit);
